Move SOLID principles doc to core
